Fix: rate 형식 int로 변경

+++ backend/src/main/java/com/rollwrite/domain/meeting/dto/MeetingCalenderResDto.java
@@ -13,7 +13,7 @@ public class MeetingCalenderResDto {
     private final String answer;
     private final int answerCnt;
     private final int participantCnt;
-    private final double rate;
+    private final int rate;
 
     @Builder
     public MeetingCalenderResDto(LocalDate day, String question, String answer, int answerCnt, int participantCnt) {
@@ -22,7 +22,7 @@ public class MeetingCalenderResDto {
         this.answer = answer;
         this.answerCnt = answerCnt;
         this.participantCnt = participantCnt;
-        this.rate = (answerCnt / (double) participantCnt) * 100;
+        this.rate = (int) ((answerCnt / (double) participantCnt) * 100);
     }
 
 }
